videoproject/requirements.txt
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.2.0
colorama==0.4.6
decorator==4.4.2
Django==4.2.3
djangorestframework==3.14.0
executing==1.2.0
ffmpeg-python==0.2.0
future==0.18.3
idna==3.4
imageio==2.25.1
imageio-ffmpeg==0.4.8
ipython==8.14.0
jedi==0.18.2
matplotlib-inline==0.1.6
moviepy==2.0.0.dev2
numpy==1.25.1
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.0
proglog==0.1.10
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
pure-eval==0.2.2
Pygments==2.15.1
pysrt==1.1.2
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
tqdm==4.65.0
traitlets==5.9.0
tzdata==2023.3
urllib3==2.0.3
wcwidth==0.2.6